requirements.txt
accelerate==1.8.1
anthropic==0.58.2
backports.tarfile==1.2.0
fqdn==1.5.1
google-genai==1.26.0
importlib-metadata==8.0.0
inspect-ai==0.3.114
isoduration==20.11.0
jaraco.collections==5.1.0
jupyter==1.1.1
langchain==0.3.26
langchain-openai==0.3.28
linkify-it-py==2.0.3
matplotlib==3.10.3
mdit-py-plugins==0.4.2
nltk==3.9.1
pandas==2.3.0
pip-chill==1.0.3
tinycss2==1.4.0
tomli==2.0.1
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
transformers==4.53.2
uri-template==1.3.0
webcolors==24.11.1
